As an end user, I require the capability to access and view detailed information pertaining to a specific client.[Delivers#186853806]

+++ models/Book.model.js
@@ -11,11 +11,11 @@ module.exports = (sequelize, Sequelize) => {
             allowNull: false,
         },
         price: {
-            type: Sequelize.DECIMAL,
+            type: Sequelize.INTEGER,
             allowNull: false,
         },
         rating: {
-            type: Sequelize.DECIMAL,
+            type: Sequelize.INTEGER,
             allowNull: false,
         },
     });
